Backend/requirements.txt
Flask==2.1.3
python-dotenv==0.19.1
APScheduler==3.10.1
celery==5.2.2
Flask-Mail==0.9.1
pandas==1.5.3
requests==2.26.0
spacy==3.5.2
numpy==1.24.3
scikit-learn==1.2.2
gensim==4.1.2
nltk==3.6.6
seaborn==0.11.2
tqdm==4.62.3
wordcloud==1.8.1
matplotlib==3.5.1
contractions==0.1.73
transformers==4.29.2
torch
redis==4.5.5
flask_cors